Add image operations and FITS utilities to PHD2 guider client
- Add get_star_image and save_image methods to Phd2Client
- Add StarImage type for PHD2 star image responses
- Add fits module with decode_base64_u16 and write_grayscale_u16_fits utilities
- Add base64 workspace dependency for image data decoding
- Update design doc to sync API signatures with implementation
- Add implementation status markers to PHD2 API reference tables
- Add comprehensive unit tests for all modules
