+++ Cargo.toml
@@ -23,3 +23,4 @@ tokio-test = "0.4"
 proptest = "1.5"
 fitsio = "0.21"
 thiserror = "2.0"
+base64 = "0.22"
