+++ services/phd2-guider/Cargo.toml
@@ -15,6 +15,8 @@ clap = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 thiserror = { workspace = true }
+base64 = { workspace = true }
+fitsio = { workspace = true }
 
 [dev-dependencies]
 tokio-test = { workspace = true }
