+++ services/phd2-guider/src/types.rs
@@ -101,6 +101,66 @@ impl fmt::Display for CalibrationTarget {
     }
 }
 
+/// Guide axis for algorithm parameter operations
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum GuideAxis {
+    /// Right Ascension axis
+    Ra,
+    /// Declination axis
+    Dec,
+}
+
+impl GuideAxis {
+    /// Get the string representation for the PHD2 API
+    pub fn to_api_string(&self) -> &'static str {
+        match self {
+            GuideAxis::Ra => "ra",
+            GuideAxis::Dec => "dec",
+        }
+    }
+}
+
+impl fmt::Display for GuideAxis {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            GuideAxis::Ra => write!(f, "RA"),
+            GuideAxis::Dec => write!(f, "Dec"),
+        }
+    }
+}
+
+/// Camera cooler status from PHD2
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CoolerStatus {
+    /// Current sensor temperature in degrees Celsius
+    pub temperature: f64,
+    /// Whether the cooler is enabled
+    #[serde(rename = "coolerOn")]
+    pub cooler_on: bool,
+    /// Target temperature setpoint in degrees Celsius (if cooler is on)
+    pub setpoint: Option<f64>,
+    /// Cooler power percentage (0-100)
+    pub power: Option<f64>,
+}
+
+/// Star image data from PHD2
+///
+/// Contains the guide star image as base64-encoded data along with metadata.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StarImage {
+    /// Frame number
+    pub frame: u64,
+    /// Image width in pixels
+    pub width: u32,
+    /// Image height in pixels
+    pub height: u32,
+    /// Star position X coordinate
+    #[serde(rename = "star_pos")]
+    pub star_pos: Option<Vec<f64>>,
+    /// Base64-encoded image data
+    pub pixels: String,
+}
+
 /// Calibration data from PHD2
 ///
 /// Contains the calibration parameters for either the mount or adaptive optics device.
@@ -129,170 +189,4 @@ pub struct CalibrationData {
     /// Declination at time of calibration (if available)
     #[serde(default)]
     pub declination: Option<f64>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_rect_creation() {
-        let rect = Rect::new(100, 200, 50, 50);
-        assert_eq!(rect.x, 100);
-        assert_eq!(rect.y, 200);
-        assert_eq!(rect.width, 50);
-        assert_eq!(rect.height, 50);
-    }
-
-    #[test]
-    fn test_rect_serialization() {
-        let rect = Rect::new(100, 200, 50, 50);
-        let json = serde_json::to_value(&rect).unwrap();
-        assert_eq!(json["x"], 100);
-        assert_eq!(json["y"], 200);
-        assert_eq!(json["width"], 50);
-        assert_eq!(json["height"], 50);
-    }
-
-    #[test]
-    fn test_profile_parsing() {
-        let json = r#"{"id":1,"name":"Default Equipment"}"#;
-        let profile: Profile = serde_json::from_str(json).unwrap();
-        assert_eq!(profile.id, 1);
-        assert_eq!(profile.name, "Default Equipment");
-    }
-
-    #[test]
-    fn test_equipment_device_parsing() {
-        let json = r#"{"name":"ZWO ASI120MM","connected":true}"#;
-        let device: EquipmentDevice = serde_json::from_str(json).unwrap();
-        assert_eq!(device.name, "ZWO ASI120MM");
-        assert!(device.connected);
-    }
-
-    #[test]
-    fn test_equipment_full_parsing() {
-        let json = r#"{
-            "camera": {"name": "ZWO ASI120MM", "connected": true},
-            "mount": {"name": "ASCOM Telescope Driver", "connected": true},
-            "aux_mount": null,
-            "AO": null,
-            "rotator": null
-        }"#;
-        let equipment: Equipment = serde_json::from_str(json).unwrap();
-
-        let camera = equipment.camera.unwrap();
-        assert_eq!(camera.name, "ZWO ASI120MM");
-        assert!(camera.connected);
-
-        let mount = equipment.mount.unwrap();
-        assert_eq!(mount.name, "ASCOM Telescope Driver");
-        assert!(mount.connected);
-
-        assert!(equipment.aux_mount.is_none());
-        assert!(equipment.ao.is_none());
-        assert!(equipment.rotator.is_none());
-    }
-
-    #[test]
-    fn test_equipment_with_ao_parsing() {
-        let json = r#"{
-            "camera": {"name": "Guide Camera", "connected": true},
-            "mount": {"name": "Mount", "connected": true},
-            "aux_mount": null,
-            "AO": {"name": "StarlightXpress AO", "connected": true},
-            "rotator": null
-        }"#;
-        let equipment: Equipment = serde_json::from_str(json).unwrap();
-
-        assert!(equipment.camera.is_some());
-        assert!(equipment.mount.is_some());
-        assert!(equipment.aux_mount.is_none());
-
-        let ao = equipment.ao.unwrap();
-        assert_eq!(ao.name, "StarlightXpress AO");
-        assert!(ao.connected);
-
-        assert!(equipment.rotator.is_none());
-    }
-
-    #[test]
-    fn test_calibration_target_get_api_string() {
-        assert_eq!(CalibrationTarget::Mount.to_get_api_string(), "Mount");
-        assert_eq!(CalibrationTarget::AO.to_get_api_string(), "AO");
-        assert_eq!(CalibrationTarget::Both.to_get_api_string(), "Mount");
-    }
-
-    #[test]
-    fn test_calibration_target_clear_api_string() {
-        assert_eq!(CalibrationTarget::Mount.to_clear_api_string(), "mount");
-        assert_eq!(CalibrationTarget::AO.to_clear_api_string(), "ao");
-        assert_eq!(CalibrationTarget::Both.to_clear_api_string(), "both");
-    }
-
-    #[test]
-    fn test_calibration_target_display() {
-        assert_eq!(format!("{}", CalibrationTarget::Mount), "Mount");
-        assert_eq!(format!("{}", CalibrationTarget::AO), "AO");
-        assert_eq!(format!("{}", CalibrationTarget::Both), "Both");
-    }
-
-    #[test]
-    fn test_calibration_data_parsing() {
-        let json = r#"{
-            "calibrated": true,
-            "xAngle": 45.0,
-            "xRate": 15.5,
-            "xParity": "+",
-            "yAngle": 135.0,
-            "yRate": 14.2,
-            "yParity": "-",
-            "declination": 30.5
-        }"#;
-        let data: CalibrationData = serde_json::from_str(json).unwrap();
-
-        assert!(data.calibrated);
-        assert_eq!(data.x_angle, 45.0);
-        assert_eq!(data.x_rate, 15.5);
-        assert_eq!(data.x_parity, "+");
-        assert_eq!(data.y_angle, 135.0);
-        assert_eq!(data.y_rate, 14.2);
-        assert_eq!(data.y_parity, "-");
-        assert_eq!(data.declination, Some(30.5));
-    }
-
-    #[test]
-    fn test_calibration_data_without_declination() {
-        let json = r#"{
-            "calibrated": true,
-            "xAngle": 45.0,
-            "xRate": 15.5,
-            "xParity": "+",
-            "yAngle": 135.0,
-            "yRate": 14.2,
-            "yParity": "-"
-        }"#;
-        let data: CalibrationData = serde_json::from_str(json).unwrap();
-
-        assert!(data.calibrated);
-        assert!(data.declination.is_none());
-    }
-
-    #[test]
-    fn test_calibration_data_not_calibrated() {
-        let json = r#"{
-            "calibrated": false,
-            "xAngle": 0.0,
-            "xRate": 0.0,
-            "xParity": "+",
-            "yAngle": 0.0,
-            "yRate": 0.0,
-            "yParity": "+"
-        }"#;
-        let data: CalibrationData = serde_json::from_str(json).unwrap();
-
-        assert!(!data.calibrated);
-        assert_eq!(data.x_angle, 0.0);
-        assert_eq!(data.x_rate, 0.0);
-    }
 }
